refactor: remove unused imports and sort used ones
