+++ src/components/ItemCard/SearchInfoCompact/styles.ts
@@ -1,0 +1,16 @@
+import { styled } from 'styled-components';
+import { COLORS } from '../../../style/GlobalStyles';
+
+export const StyledItemCardCompactContainer = styled.div`
+    display: flex;
+    background-color: ${COLORS.lightGray};
+    margin: 16px;
+    box-shadow: 2px 4px 4px 0 rgba(0, 0, 0, 0.2);
+    flex-direction: column;
+    max-width: 500px;
+    position: relative;
+`;
+
+export const StyledCompactInfoP = styled.div`
+    width: 100%;
+`;
