+++ src/pages/listDetails/phone/AddMore.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useDebounce } from 'usehooks-ts';
 import { GlobalSpinner } from '../../../components/GlobalSpinner/GlobalSpinner';
-import SearchResultCardCompact from '../../../components/ItemCard/SearchInfoCompact';
+import SearchResultCardCompact from '../../../components/ItemCard/SearchInfoCompact/SearchInfoCompact';
 
 import SearchBar from '../../../components/SearchBar/SearchBar';
 import { useSnackBar } from '../../../hooks';
